Implemented initial support for time markers.
Time markers are a special type of text that internally save a
specific moment in time compared to the start of a score (time is 0 at
that point). Unlike other types of text, time markers do not move with
bars/beats/etc, but instead move to their correct location to ensure
they happen at the same specific moment in time.

In addition to adding support for this new text type, I've also added
a visual affordance that spans all staves to clearly show where the
moments in time occur across all parts.